Move all component to Layout routing
