--- frontend/src/components/account/Account.js
@@ -1,13 +1,0 @@
-import React from "react";
-import { Card } from "antd";
-
-const Account = ({ account, certificates }) => {
-  return (
-    <Card title={`Account: ${account.account_name}`}>
-      <p>Number of Certificates: {certificates.length}</p>
-      <p>Number of Devices: {account.devices.length}</p>
-    </Card>
-  );
-};
-
-export default Account;
