+++ frontend/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import Root from './Root'; 
+import App from './App'; 
 import 'antd/dist/antd.js';
 import "./assets/styles/global.css";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<Root />);
+root.render(<App />);
